# [SPARK-7673] [SQL] WIP: HadoopFsRelation and ParquetRelation2 performance optimizations